feat: encode branches to work with slashes
This currently only works from the form from the home page. The redirect cannot determine what part is the branch. The GitHub API does not expose an endpoint to reliably provide a list of all branches. Technically it would be possible to iterate over every possible branch name from the given path in the URL and send a request and check for a 404 or 200 status, but this is currently not implemented.